extracted url building to a helper class
