+++ collectors/scm/gitlab/src/main/java/com/capitalone/dashboard/collecteur/DefaultGitlabGitClient.java
@@ -1,23 +1,15 @@
 package com.capitalone.dashboard.collecteur;
 
-import java.net.MalformedURLException;
 import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
 import java.util.List;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.joda.time.DateTime;
@@ -30,7 +22,6 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestOperations;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import com.capitalone.dashboard.model.Commit;
 import com.capitalone.dashboard.model.GitlabGitRepo;
@@ -41,35 +32,25 @@ import com.capitalone.dashboard.util.Supplier;
  */
 
 @Component
-@SuppressWarnings({"PMD"})
 public class DefaultGitlabGitClient implements  GitlabGitClient {
 
-	private final GitlabSettings gitlabSettings;
+	private final GitlabUrlUtility gitlabUrlUtility;
 
     private static final Log LOG = LogFactory.getLog(DefaultGitlabGitClient.class);
 
     private final RestOperations restOperations;
-
-	private static final String PROTOCOL = "https";
-    private static final String SEGMENT_API = "/api/v3/projects/";
-	private static final String COMMITS_API = "/repository/commits/";
-	private static final String PRIVATE_TOKEN_QUERY_PARAM_KEY = "private_token";
-	private static final String DATE_QUERY_PARAM_KEY = "since";
-	private static final String BRANCH_QUERY_PARAM_KEY = "ref_name";
-    private static final String PUBLIC_GITLAB_HOST_NAME = "gitlab.company.com";
-	private static final int FIRST_RUN_HISTORY_DEFAULT = 14;
     
     @Autowired
-    public DefaultGitlabGitClient(GitlabSettings gitlabSettings,
+    public DefaultGitlabGitClient(GitlabUrlUtility gitlabUrlUtility,
                                        Supplier<RestOperations> restOperationsSupplier) {
-        this.gitlabSettings = gitlabSettings;
+        this.gitlabUrlUtility = gitlabUrlUtility;
         this.restOperations = restOperationsSupplier.get();
     }
 
     @Override
 	public List<Commit> getCommits(GitlabGitRepo repo, boolean firstRun) {
         List<Commit> commits = new ArrayList<>();
-		String apiUrl = buildApiUrl(repo, firstRun);
+		URI apiUrl = gitlabUrlUtility.buildApiUrl(repo, firstRun);
 
 		ResponseEntity<String> response = makeRestCall(apiUrl);
 		JSONArray jsonArray = paresAsArray(response);
@@ -81,80 +62,7 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
         return commits;
     }
 
-	private String buildApiUrl(GitlabGitRepo repo, boolean firstRun) {
-		String repoUrl = repo.getRepoUrl();
-        if (repoUrl.endsWith(".git")) {
-            repoUrl = repoUrl.substring(0, repoUrl.lastIndexOf(".git"));
-        }
-        
-		String repoName = getRepoName(repoUrl);
-		String host = getRepoHost();
-		String date = getDateForCommits(repo, firstRun);
-
-		UriComponentsBuilder builder = UriComponentsBuilder.newInstance();
-		String fullUrl = builder.scheme(PROTOCOL)
-				.host(host)
-				.path(SEGMENT_API)
-				.path(repoName)
-				.path(COMMITS_API)
-				.queryParam(BRANCH_QUERY_PARAM_KEY, repo.getBranch())
-				.queryParam(DATE_QUERY_PARAM_KEY, date)
-				.queryParam(PRIVATE_TOKEN_QUERY_PARAM_KEY, gitlabSettings.getApiToken())
-				.build(true)
-				.toUriString();
-
-		return fullUrl;
-    }
-
-	private String getRepoHost() {
-		String providedGitLabHost = gitlabSettings.getHost();
-		String apiHost;
-		if (StringUtils.isBlank(providedGitLabHost)) {
-			apiHost = PUBLIC_GITLAB_HOST_NAME;
-		} else {
-			apiHost = providedGitLabHost;
-		}
-		return apiHost;
-	}
-
-	private String getRepoName(String repoUrl) {
-		String repoName = "";
-		try {
-			URL url = new URL(repoUrl);
-			repoName = url.getFile();
-		} catch (MalformedURLException e) {
-			LOG.error(e.getMessage());
-		}
-
-		repoName = repoName.substring(repoName.indexOf("/") + 1, repoName.length());
-		repoName = repoName.replace("/", "%2F");
-		return repoName;
-	}
-
-	private String getDateForCommits(GitlabGitRepo repo, boolean firstRun) {
-		Date dt;
-		if (firstRun) {
-			int firstRunDaysHistory = gitlabSettings.getFirstRunHistoryDays();
-			if (firstRunDaysHistory > 0) {
-				dt = getDate(new Date(), -firstRunDaysHistory, 0);
-			} else {
-				dt = getDate(new Date(), -FIRST_RUN_HISTORY_DEFAULT, 0);
-			}
-		} else {
-			dt = getDate(new Date(repo.getLastUpdated()), 0, -10);
-		}
-		DateFormat df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'");
-		String thisMoment = df.format(dt);
-		return thisMoment;
-	}
-
-	private Date getDate(Date dateInstance, int offsetDays, int offsetMinutes) {
-		Calendar cal = Calendar.getInstance();
-		cal.setTime(dateInstance);
-		cal.add(Calendar.DATE, offsetDays);
-		cal.add(Calendar.MINUTE, offsetMinutes);
-		return cal.getTime();
-	}
+	
 
 	private Commit buildCommit(JSONObject jsonObject, String repoUrl, String repoBranch) {
 		String author = str(jsonObject, "author_name");
@@ -175,16 +83,9 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
 		return commit;
 	}
 
-	private ResponseEntity<String> makeRestCall(String url) {
+	private ResponseEntity<String> makeRestCall(URI url) {
 		trustSelfSignedSSL();
-		URI uri = null;
-		try {
-			uri = new URI(url);
-		} catch (URISyntaxException e) {
-			LOG.error(e.getMessage());
-		}
-
-		return restOperations.exchange(uri, HttpMethod.GET, null, String.class);
+		return restOperations.exchange(url, HttpMethod.GET, null, String.class);
 
 	}
